+++ pdf/test/all_tests.dart
@@ -38,6 +38,7 @@ import 'widget_container_test.dart' as widget_container;
 import 'widget_flex_test.dart' as widget_flex;
 import 'widget_form_test.dart' as widget_form;
 import 'widget_grid_view_test.dart' as widget_grid_view;
+import 'widget_icon_test.dart' as widget_icon;
 import 'widget_multipage_test.dart' as widget_multipage;
 import 'widget_opacity_test.dart' as widget_opacity;
 import 'widget_outline_test.dart' as widget_outline;
@@ -72,6 +73,7 @@ void main() {
   widget_flex.main();
   widget_form.main();
   widget_grid_view.main();
+  widget_icon.main();
   widget_multipage.main();
   widget_opacity.main();
   widget_outline.main();